crypt_scan/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
backports.ssl-match-hostname==3.7.0.1
chardet==4.0.0
colored==1.4.2
gevent==21.1.2
greenlet==1.0.0
idna==3.1
multidict==5.1.0
python-dotenv==0.15.0
six==1.15.0
typing-extensions==3.7.4.3
websocket==0.2.1
websocket-client==0.58.0
websockets==8.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
